AsyncElement: Prototype for AsyncElement support
This is an example impl I thought up for Async Element.  In this impl
the element implements both Future and Stream.  The Future poll is
the one that is handed to the runtime, and the Stream is what the
element after this element uses to pop packets from the queue.
I think this 'hack' might work pretty well since it ends up pretty
clean, and theoretically we should be able to have two impls of
poll on the one link element.

There is no test code for this yet. There is one bug in that the
Future poll returns NotReady when it is unable to do more work,
and from my reading of Tokio, I do not think it will be woken up
again, resulting in a hang, unless we implement some sort of
reactor that will watch the input stream.